refactor: get database definition
